Drop the `build-time` dependency
The `build-time` dependency has a large dependency tree. But we don't really
need an ISO-formatted timestamp. We just need to do a visual equality check on
the time. Therefore, seconds since the Unix epoch is a reasonable alternative.
So, instead of the `build-time` crate, just calculate the time of the build in
a build script and reference that time as an environment variable.

Signed-off-by: Nathaniel McCallum <nathaniel@profian.com>