project: shuffle crates around, bump version
Having the main crate as the binary has the unfortunate effect of not
being documented at docs.rs. So now the runtime crate is named 'peginator'
and the CLI crate is separate.

This has the nice side-effect of people not having to change their current
code, as PegParser will still come from the 'peginator' crate.